==> preprocessing/3c-FIDDLE2-discretize-labs.ipynb <==
import joblib
import pandas as pd
import numpy as np
import sparse
import json

### Labs

df_time_series = joblib.load('sample_output/out_labs/sdf.joblib')
discretization_bins = json.load(open('metadata/out_labs/discretization.json', 'r'))
feature_names = json.load(open('metadata/out_labs/X_all.feature_names.json', 'r'))
len(discretization_bins)

cols_useful = list(discretization_bins.keys())

df_time_series = df_time_series[cols_useful]

df_time_series

ddf = df_time_series.sparse.to_dense()

from steps_2 import *
from helpers import *

df = ddf

df_by_cols = [df[col] for col in df.columns]

pool = multiprocessing.Pool(multiprocessing.cpu_count())
out_0 = list(tqdm(pool.imap_unordered(
    smart_qcut_bins,
    [(df[col], discretization_bins[col]) for col in df.columns]), total=len(df.columns)
))
cols_data, dis_bins = zip(*out_0)
out = list(tqdm(pool.imap_unordered(dummify, cols_data), total=len(df.columns)
))
pool.close()
pool.join()

out_0 = [smart_qcut_bins(x, discretization_bins[x.name]) for x in tqdm(df_by_cols)]
cols_data, dis_bins = zip(*out_0)
out = [dummify(z) for z in tqdm(cols_data)]

time_invariant_features = pd.concat(out, axis=1).sort_index(axis=1)
feature_names_all = time_invariant_features.columns.values
sdf = time_invariant_features.astype(pd.SparseDtype(int, fill_value=0))
s_ = sparse.COO(sdf.sparse.to_coo())

df_features = pd.concat(out, axis=1).sort_index(axis=1)

# Drop any values='missing' for now
df_features = df_features.loc[:, [col for col in df_features.columns if 'missing' not in col]]

# Check same features
set(df_features.columns) <= set(feature_names)

df_features = df_features.reindex(columns=feature_names, fill_value=0)

df_features

sdf = df_features.astype(pd.SparseDtype(int, fill_value=0))
X_all = sparse.COO(sdf.sparse.to_coo())

sparse.save_npz('sample_output/out_labs/X_all.npz', X_all)
df_time_series[[]].to_csv('sample_output/out_labs/X_all.IDs.csv')

X_all.density
